Subplot appends add_<graphic_name> methods, so now available in GridPlot
